Replace implementation-dependent type long int

+++ CRTL/impl/win/main.cpp
@@ -29,7 +29,7 @@
 
 extern int main (int argc, char* argv [], char** envp);
 
-extern "C" unsigned long mainCRTStartup (void)
+extern "C" uint32_t mainCRTStartup (void)
 {
 	if (!CRTL::initialize ())
 		return 255;
